src/web-server/pyproject.toml
[project]
dependencies = [
  "uvicorn[standard]>=0.30.3",
  "fastapi[standard]>=0.111.1",
  "aiohttp>=3.9.5",
  "opentelemetry-distro>=0.48b0",
  "opentelemetry-instrumentation-fastapi>=0.48b0",
  "opentelemetry-exporter-otlp-proto-http>=1.27.0",
  "opentelemetry-exporter-otlp>=1.27.0",
  "opentelemetry-instrumentation-aiohttp-client>=0.48b0",
  # "grpcio>=1.67.1",
  # "grpcio-tools>=1.67.1",
]
name = "webserver"
version = "0.1.0"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["*.py"]
